+++ functions/create_draft/handler_test.ts
@@ -5,8 +5,8 @@ import {
   assertFalse,
 } from "https://deno.land/std@0.153.0/testing/asserts.ts";
 
-import createDraft from "./mod.ts";
-import { CREATE_DRAFT_FUNCTION_CALLBACK_ID } from "./definition.ts";
+import createDraft from "./handler.ts";
+import { CREATE_DRAFT_FUNCTION_CALLBACK_ID } from "./mod.ts";
 
 const { createContext } = SlackFunctionTester(
   CREATE_DRAFT_FUNCTION_CALLBACK_ID,
@@ -15,7 +15,7 @@ const { createContext } = SlackFunctionTester(
 // Replace global this.fetch with mocked copy
 mf.install();
 
-// Setup inputs based on function inputs and outputs in "./definition.ts"
+// Setup inputs based on function inputs and outputs in "./mod.ts"
 const inputs = {
   created_by: "dummyUserId",
   message: "The content of the announcement",
